Backend/requirements.txt
langchain
streamlit
langchain-google-genai
pillow
pdfx
pyGithub
fastapi
uvicorn
python-multipart